Small name refactor to simplify.
