+++ research/flamby/fed_heart_disease/fenda/client.py
@@ -15,7 +15,7 @@ from torch.optim import Optimizer
 from torch.utils.data import DataLoader
 
 from fl4health.checkpointing.checkpointer import BestLossTorchCheckpointer, LatestTorchCheckpointer
-from fl4health.checkpointing.client_side_module import ClientSideCheckpointModule
+from fl4health.checkpointing.client_module import ClientCheckpointModule
 from fl4health.clients.fenda_client import FendaClient
 from fl4health.utils.losses import LossMeterType
 from fl4health.utils.metrics import Accuracy, Metric
@@ -33,7 +33,7 @@ class FedHeartDiseaseFendaClient(FendaClient):
         client_number: int,
         learning_rate: float,
         loss_meter_type: LossMeterType = LossMeterType.AVERAGE,
-        checkpointer: Optional[ClientSideCheckpointModule] = None,
+        checkpointer: Optional[ClientCheckpointModule] = None,
         cos_sim_activate: bool = False,
         contrastive_activate: bool = False,
         extra_loss_weights: Optional[float] = None,
@@ -155,7 +155,7 @@ if __name__ == "__main__":
         if federated_checkpointing
         else LatestTorchCheckpointer(checkpoint_dir, checkpoint_name)
     )
-    checkpointer = ClientSideCheckpointModule(post_aggregation=post_aggregation_checkpointer)
+    checkpointer = ClientCheckpointModule(post_aggregation=post_aggregation_checkpointer)
 
     client = FedHeartDiseaseFendaClient(
         data_path=Path(args.dataset_dir),
